Bunch of pending updates
Most of these changes are not final but I never have any time to review
and update stuff properly. Seem to work for now anyway.
Main change is adding pipenv "repo": two directories with pipfiles
and a script and installs requirements and copies binary shims from
virtualenvs to ~/.local/bin (it doesn't work quite like I'd want it to,
I have to remove some of the shims manually for now, but definitely beats
pip install --user)